requirements.txt
streamlit
google-genai
Pillow
python-dotenv
st-gsheets-connection
pandas
google-api-python-client 
google-auth
pillow